Driver-Drowsiness-Detection-testing/requirements.txt
numpy
opencv-python
opencv-contrib-python
mediapipe
requests
RPi.GPIO
smbus2
